TST/REF: misplaced .xs tests (#37456)
