+++ pandas/tests/series/indexing/test_getitem.py
@@ -278,3 +278,15 @@ def test_getitem_ndim_deprecated():
     s = Series([0, 1])
     with tm.assert_produces_warning(FutureWarning):
         s[:, None]
+
+
+def test_getitem_multilevel_scalar_slice_not_implemented(
+    multiindex_year_month_day_dataframe_random_data,
+):
+    # not implementing this for now
+    df = multiindex_year_month_day_dataframe_random_data
+    ser = df["A"]
+
+    msg = r"\(2000, slice\(3, 4, None\)\)"
+    with pytest.raises(TypeError, match=msg):
+        ser[2000, 3:4]
